% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.R
\name{.marxan_alpha_parameters}
\alias{.marxan_alpha_parameters}
\title{Marxan feature alpha irreplaceability parameters}
\usage{
.marxan_alpha_parameters(spec, puvspr, spid)
}
\arguments{
\item{spec}{data.frame - The spec.dat as a data.frame.}

\item{puvspr}{data.frame - The puvspr.dat as a data.frame.}

\item{spid}{integer - The species ID.}
}
\value{
A list
}
\description{
Returns a feature's target and global representation
}
\examples{
\dontrun{
.marxan_alpha_parameters(spec_dataframe, puvsp_dataframe, 17)
}
}
\author{
Daniele Baisero, \email{daniele.baisero@gmail.com}
}
\keyword{internal}
