% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.R
\name{.marxan_input_parameter}
\alias{.marxan_input_parameter}
\title{Marxan input.dat parameter reader}
\usage{
.marxan_input_parameter(input, parameter)
}
\arguments{
\item{input}{string - The address of a Marxan input.dat file.}

\item{parameter}{string - The name of the parameter to read.}
}
\value{
A string.
}
\description{
Reads a marxan input.dat file and returns a single parameter.
}
\examples{
\dontrun{
.marxan_read_input('/data/marxan/analysis01/input.dat', 'INPUTDIR')
}
}
\author{
Daniele Baisero, \email{daniele.baisero@gmail.com}
}
\keyword{internal}
