% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.R
\name{marxan_gammas}
\alias{marxan_gammas}
\title{Marxan: Gamma irreplaceabilities}
\usage{
marxan_gammas(input, triage = FALSE)
}
\arguments{
\item{input}{string - The address of the input.dat file.}

\item{triage}{logical - Should features with unachievable targets be given an
irreplaceability of 0? See \code{\link{gamma}}.}
}
\value{
A data.frame
}
\description{
Given a valid Marxan input.dat file, returns a copy of the PUNAME file
(pu.dat), with an additional column containing the planning unit's gamma of
irreplaceability.
}
\examples{
\dontrun{
marxan_gammas('/data/marxan/analysis01/input.dat')
marxan_gammas('C:\data\marxan\analysis01\input.dat')
}
}
\author{
Daniele Baisero, \email{daniele.baisero@gmail.com}
}
